Bux fixes in chooseAvatar and added advances in joinChat modal

+++ src/main/resources/static/src/components/register/modals/item-chooser/ModalIconChooser.jsx
@@ -1,13 +1,13 @@
-import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
-import '../Register.css';
+import '../../Register.css';
 import './ModalIconChooser.css';
-import { imageLinks } from '../../../services/avatarsLinks.js';
+import { imageLinks } from '../../../../services/avatarsLinks.js';
 import { useState } from 'react';
 import ItemAvatar from './ItemAvatar.jsx';
 
-const ModalIconChooser = ({ show, handleClose }) => {
+const ModalIconChooser = ({ showModalIconChooser, handleCloseModalIconChooser }) => {
 
+  const [iconPrevChoosed, setIconPrevChoosed] = useState(localStorage.getItem('avatarImg'));
   const [iconChoosed, setIconChoosed] = useState(localStorage.getItem('avatarImg'));
   const [itemActiveIndex, setItemActiveIndex] = useState(null);
 
@@ -16,13 +16,19 @@ const ModalIconChooser = ({ show, handleClose }) => {
     setIconChoosed(urlIcon);
   }
 
-  const handleConfirmItemChoosed = () => {
-    localStorage.setItem('avatarImg', iconChoosed)
-    return handleClose(iconChoosed);
+  const handleConfirmItemChoosed = (isConfirmButton) => {
+    if(isConfirmButton){
+      localStorage.setItem('avatarImg', iconChoosed)
+      return handleCloseModalIconChooser(iconChoosed);
+    }else{
+      return handleCloseModalIconChooser(iconPrevChoosed);
+    }
+    
   }
 
   return (
-    <Modal show={show} >
+    <Modal show={showModalIconChooser} 
+           onHide={()=> handleConfirmItemChoosed(false)}>
       <Modal.Header closeButton>
         <Modal.Title>Choose Avatar</Modal.Title>
       </Modal.Header>
@@ -32,8 +38,9 @@ const ModalIconChooser = ({ show, handleClose }) => {
             <ul>
               {imageLinks.map((url, i) => (
                 <li
+                  key={i}
                   className={`list-item-item ${i === itemActiveIndex ? 'item-active' : ''}`}
-                  onClick={() => { setItemActive(i) }}
+                  onClick={() => { setItemActiveIndex(i) }}
                 >
                   <ItemAvatar url={url} i={i} handleItemChoosed={handleItemChoosed} />
                 </li>
@@ -41,18 +48,13 @@ const ModalIconChooser = ({ show, handleClose }) => {
               }
 
             </ul>
-
-            {
-              /*                 2jl8plylc
-              */
-            }
           </div>
         </div>
 
 
       </Modal.Body>
       <Modal.Footer>
-        <button className='button button-chooser' onClick={handleConfirmItemChoosed}>
+        <button className='button button-chooser' onClick={()=>handleConfirmItemChoosed(true)}>
           Confirm
         </button>
       </Modal.Footer>
